+++ src/api/endpoints.ts
@@ -17,3 +17,23 @@ export const getProjects: Endpoint = {
   method: 'GET',
   url: `${CORE.rangerBaseApi}/projects`
 }
+
+export const getIncidents: Endpoint = {
+  method: 'GET',
+  url: `${CORE.rangerBaseApi}/incidents`
+}
+
+export const getResponse: Endpoint = {
+  method: 'GET',
+  url: `${CORE.rangerBaseApi}/responses`
+}
+
+export const getAssets: Endpoint = {
+  method: 'GET',
+  url: `${CORE.rangerBaseApi}/assets`
+}
+
+export const closeIncident: Endpoint = {
+  method: 'PATCH',
+  url: `${CORE.rangerBaseApi}/incidents`
+}