+++ src/components/project-selector/project-selector.ts
@@ -28,7 +28,7 @@ export default class ProjectSelectorComponent extends Vue {
 
   async confirmedSelectedProject (): Promise<void> {
     const newProjectId = this.selectedProject?.id
-    if (newProjectId !== undefined) void this.$router.push({ name: ROUTES_NAME.guardians, params: { projectId: newProjectId } })
+    if (newProjectId !== undefined) void this.$router.push({ name: ROUTES_NAME.incidents, params: { projectId: newProjectId } })
     this.closeProjectSelector()
   }