+++ src/models/Image.ts
@@ -1,0 +1,5 @@
+export interface RawImageItem {
+  src: string,
+  width?: number,
+  heigth?: number
+}