+++ src/models/Stream.ts
@@ -1,7 +1,7 @@
-// TODO: Update after connect to /stream endpoint
 export interface Stream {
   id: string
   name: string
+  timezone: string
   latitude?: number
   longitude?: number
 }