+++ src/pages/guardians/guardians.vue
@@ -1,2 +1,0 @@
-<template src="./guardians.html" />
-<script src="./guardians.ts" lang="ts" />