+++ src/pages/incidents/incidents.vue
@@ -1,0 +1,2 @@
+<template src="./incidents.html" />
+<script src="./incidents.ts" lang="ts" />