+++ src/pages/index.ts
@@ -1,7 +1,5 @@
 export { default as ErrorPage } from './error/error.vue'
 export { default as IndexPage } from './index/index.vue'
 export { default as RootPage } from './root/root.vue'
-export { default as SpeciesRichnessPage } from './species-richness/species-richness.vue'
-
-export { default as GuardiansPage } from './guardians/guardians.vue'
-export { default as ReportPage } from './report/report.vue'
+export { default as IncidentsPage } from './incidents/incidents.vue'
+export { default as IncidentPage } from './incident/incident.vue'