+++ src/pages/report/report.vue
@@ -1,2 +1,0 @@
-<template src="./report.html" />
-<script src="./report.ts" lang="ts" />