+++ src/services/streams.service.ts
@@ -1,5 +1,4 @@
 import * as Endpoints from '@/api/endpoints'
-import rawStreams from '@/api/raw-site-01-07-apr-2021.json'
 import { StreamModels } from '@/models'
 import ApiClient from './api.service'
 
@@ -17,16 +16,4 @@ export async function getStreams (projectsId: any): Promise<StreamModels.Stream[
   } catch (e) {
     return await Promise.reject(e)
   }
-}
-
-export function getMockupStreams (): StreamModels.Stream[] {
-  return rawStreams.map(s => {
-    const { name, latitude, longitude } = s
-    return {
-      id: s.site_id,
-      name,
-      latitude: Number(latitude),
-      longitude: Number(longitude)
-    }
-  })
 }