+++ src/stores/actions.ts
@@ -1,5 +1,6 @@
 export const root = {
   updateAuth: 'updateAuth',
   updateUser: 'updateUser',
+  updateStreams: 'updateStreams',
   updateSelectedProject: 'updateSelectedProject'
 }