+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## 0.0.1 (2021-??-??)
 
-* **setup:** CE-1446 The Ranger Track modal window is created ([CE-1446](https://jira.rfcx.org/browse/CE-1446))
-* **setup:** CE-1429 Implement UI for incidents and incident detail pages ([CE-1429](https://jira.rfcx.org/browse/CE-1429))
-* **setup:** CE-1363 Create mockup of Incident Center homepage in HTML/WindiCSS ([CE-1363](https://jira.rfcx.org/browse/CE-1261))
+* **setup:** Display three events in the list of incidents([CE-1448](https://jira.rfcx.org/browse/CE-1448))
+* **setup:** The Ranger Track modal window is created ([CE-1446](https://jira.rfcx.org/browse/CE-1446))
+* **setup:** Implement UI for incidents and incident detail pages ([CE-1429](https://jira.rfcx.org/browse/CE-1429))
+* **setup:** Create mockup of Incident Center homepage in HTML/WindiCSS ([CE-1363](https://jira.rfcx.org/browse/CE-1261))